requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-allauth==0.41.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-next-prev==1.0.1
django-tinymce4-lite==1.7.5
gunicorn==20.0.4
idna==2.9
isort==4.3.21
jsmin==2.2.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
oauthlib==3.1.0
pep8==1.7.1
Pillow==7.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.11.2
